Create NotificationPanelButton component
