+++ src/components/views/spaces/SpacePanel.tsx
@@ -60,7 +60,7 @@ import SettingsStore from "../../../settings/SettingsStore";
 import { SettingLevel } from "../../../settings/SettingLevel";
 import UIStore from "../../../stores/UIStore";
 import QuickSettingsButton from "./QuickSettingsButton";
-import { useSettingValue } from "../../../hooks/useSettings";
+import { useFeatureEnabled, useSettingValue } from "../../../hooks/useSettings";
 import UserMenu from "../../structures/UserMenu";
 import IndicatorScrollbar from "../../structures/IndicatorScrollbar";
 import { IS_MAC, Key } from "../../../Keyboard";
@@ -72,14 +72,7 @@ import { NotificationState } from "../../../stores/notifications/NotificationSta
 import { ALTERNATE_KEY_NAME } from "../../../accessibility/KeyboardShortcuts";
 import { shouldShowComponent } from "../../../customisations/helpers/UIComponents";
 import { UIComponent } from "../../../settings/UIFeature";
-import RightPanelStore from "../../../stores/right-panel/RightPanelStore";
-import { useGlobalNotificationState } from "../../../hooks/useGlobalNotification";
-import { NotificationColor } from "../../../stores/notifications/NotificationColor";
-import { RightPanelPhases } from "../../../stores/right-panel/RightPanelStorePhases";
-import { UnreadIndicator } from "../right_panel/LegacyRoomHeaderButtons";
-import { Icon as NotificationIcon } from "../../../../res/img/element-icons/notifications.svg";
-import TooltipTarget from "../elements/TooltipTarget";
-import { Alignment } from "../elements/Tooltip";
+import { NotificationPanelButton } from "./NotificationPanelButton";
 
 const useSpaces = (): [Room[], MetaSpace[], Room[], SpaceKey] => {
     const invites = useEventEmitterState<Room[]>(SpaceStore.instance, UPDATE_INVITED_SPACES, () => {
@@ -352,7 +345,7 @@ const SpacePanel: React.FC = () => {
         }
     });
 
-    const summarizedNotificationState = useGlobalNotificationState();
+    const isNotifPanelEnabled = useFeatureEnabled("feature_notifications_panel");
 
     return (
         <RovingTabIndexProvider handleHomeEnd handleUpDown={!dragging}>
@@ -409,22 +402,7 @@ const SpacePanel: React.FC = () => {
                             )}
                         </Droppable>
 
-                        {SettingsStore.getValue("feature_video_rooms") && (
-                            <TooltipTarget label={_t("Notifications")} alignment={Alignment.Right}>
-                                <div
-                                    className="mx_SpacePanel_button"
-                                    title={_t("Notifications")}
-                                    onClick={() => {
-                                        RightPanelStore.instance.setCard({ phase: RightPanelPhases.NotificationPanel });
-                                    }}
-                                >
-                                    <NotificationIcon />
-                                    {summarizedNotificationState.color === NotificationColor.Red ? (
-                                        <UnreadIndicator color={summarizedNotificationState.color} />
-                                    ) : null}
-                                </div>
-                            </TooltipTarget>
-                        )}
+                        {isNotifPanelEnabled && <NotificationPanelButton />}
                         <QuickSettingsButton className="mx_SpacePanel_button" isPanelCollapsed={isPanelCollapsed} />
                     </div>
                 </DragDropContext>
